+++ srcgen/web-entity/%libs.platformFolder%/src/components/entities/%entity.name.kebab%/%entity.list.name.kebab%-grid-modal/%entity.list.name.kebab%-grid-modal.module.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { ModuleWithProviders, NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { TranslateModule } from '@ngx-translate/core';
-import { EntityGridModalModule } from '@rucken/core';
+import { EntityGridModalModule } from '@rucken/web';
 import { <%=entity.list.name.camel%>GridModule } from '../<%=entity.list.name.kebab%>-grid/<%=entity.list.name.kebab%>-grid.module';
 import { <%=entity.list.name.camel%>GridModalComponent } from './<%=entity.list.name.kebab%>-grid-modal.component';