+++ srcgen/web-entity/%libs.platformFolder%/src/components/entities/%entity.name.kebab%/%entity.name.kebab%-input/%entity.name.kebab%-input.module.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { ModuleWithProviders, NgModule } from '@angular/core';
 import { TranslateModule } from '@ngx-translate/core';
 import { ModalModule } from 'ngx-bootstrap/modal';
-import { EntityInputModule } from '@rucken/core';
+import { EntityInputModule } from '@rucken/web';
 import { <%=entity.list.name.camel%>GridModalModule } from '../<%=entity.list.name.kebab%>-grid-modal/<%=entity.list.name.kebab%>-grid-modal.module';
 import { <%=entity.name.camel%>InputComponent } from './<%=entity.name.kebab%>-input.component';