+++ srcgen/web-entity/%libs.platformFolder%/src/components/entities/%entity.name.kebab%/%entity.name.kebab%-select/%entity.name.kebab%-select.module.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { ModuleWithProviders, NgModule } from '@angular/core';
 import { TranslateModule } from '@ngx-translate/core';
 import { ModalModule } from 'ngx-bootstrap/modal';
-import { EntitySelectModule } from '@rucken/core';
+import { EntitySelectModule } from '@rucken/web';
 import { <%=entity.list.name.camel%>GridModalModule } from '../<%=entity.list.name.kebab%>-grid-modal/<%=entity.list.name.kebab%>-grid-modal.module';
 import { <%=entity.name.camel%>SelectComponent } from './<%=entity.name.kebab%>-select.component';